// src/components/hashtag/hashtag.types.d.ts
export interface Settings {
  background_color: string;
  name: string;
  title: string;
  subtitle: string;
  link_1: string;
  image_1: string | null;
  link_2: string;
  image_2: string | null;
  link_3: string;
  image_3: string | null;
  link_4: string;
  image_4: string | null;
  link_5: string;
  image_5: string | null;
  link_6: string;
  image_6: string | null;
}